Flashing LED test board: Update Upverter URL in README.md
Upverter was having problems; I couldn't open the project in the editor, it just kept saying server connection error or something like that. I got around it by forking the project, deleting the original and then setting the metadata on the forked project (name & description) to match the original. This new URL points to the forked project but that's the "real" one now becasue the original is deleted.